+++ CoffeLand/CoffeeLand_UI/Areas/Admin/Controllers/CoffeeController.cs
@@ -6,7 +6,6 @@ using System.Linq;
 using System.Net;
 using System.Web;
 using System.Web.Mvc;
-using CoffeeLand_BLL.Repository.Abstract;
 using CoffeeLand_BLL.Repository.Concrete;
 using CoffeeLand_DAL;
 using CoffeeLand_DATA.Classes;
@@ -15,36 +14,77 @@ namespace CoffeeLand_UI.Areas.Admin.Controllers
 {
     public class CoffeeController : Controller
     {
-        CoffeeConcrete _coffeeConcrete;
-        CategoryConcrete _categoryConcrete;
-        ExtraMaterialsConcrete _extraMaterialsConcrete;
+		CoffeeConcrete _coffeeConcrete;
+		CategoryConcrete _categoryConcrete;
+		ExtraMaterialsConcrete _extraMaterialsConcrete;
 
-        public CoffeeController()
-        {
-            _coffeeConcrete = new CoffeeConcrete();
-            _categoryConcrete = new CategoryConcrete();
-            _extraMaterialsConcrete = new ExtraMaterialsConcrete();
-        }
+		public CoffeeController()
+		{
+			_coffeeConcrete = new CoffeeConcrete();
+			_categoryConcrete = new CategoryConcrete();
+			_extraMaterialsConcrete = new ExtraMaterialsConcrete();
+		}
 
-        // GET: Admin/Coffee
-        public ActionResult Index()
+		// GET: Admin/Coffee
+		public ActionResult Index()
         {
-            return View(_coffeeConcrete._coffeeRepository.GetAll());
+			Customer customer = Session["OnlineKullanici"] as Customer;
+
+			if (customer == null)
+			{
+				return Redirect("/Login/Login");
+			}
+			else if (customer.AuthorizationID == 1 || customer.AuthorizationID == 2)
+			{
+				return View(_coffeeConcrete._coffeeRepository.GetAll());
+			}
+			else
+			{
+				return Redirect("/Coffee/Coffees");
+			}
         }
 
         // GET: Admin/Coffee/Details/5
         public ActionResult Details(int id)
         {
-            Coffee coffee = _coffeeConcrete._coffeeRepository.GetById(id);
-            return View(coffee);
+			Customer customer = Session["OnlineKullanici"] as Customer;
+
+			if (customer == null)
+			{
+				return Redirect("/Login/Login");
+			}
+			else if (customer.AuthorizationID == 1 || customer.AuthorizationID == 2)
+			{
+				Coffee coffee = _coffeeConcrete._coffeeRepository.GetById(id);
+				return View(coffee);
+			}
+			else
+			{
+				return Redirect("/Coffee/Coffees");
+			}
+			
         }
 
         // GET: Admin/Coffee/Create
         public ActionResult Create()
         {
-            ViewBag.CategoryID = new SelectList(_categoryConcrete._categoryRepository.GetEntity(), "ID", "CategoryName");
-            ViewBag.ExtraMaterialsID = new SelectList(_extraMaterialsConcrete._extraMaterialRepository.GetEntity(), "ID", "Name");
-            return View();
+			Customer customer = Session["OnlineKullanici"] as Customer;
+
+			if (customer == null)
+			{
+				return Redirect("/Login/Login");
+			}
+			else if (customer.AuthorizationID == 1 || customer.AuthorizationID == 2)
+			{
+				ViewBag.CategoryID = new SelectList(_categoryConcrete._categoryRepository.GetEntity(), "ID", "CategoryName");
+				ViewBag.ExtraMaterialsID = new SelectList(_categoryConcrete._categoryRepository.GetEntity(), "ID", "Name");
+				return View();
+			}
+			else
+			{
+				return Redirect("/Coffee/Coffees");
+			}
+			
         }
 
         // POST: Admin/Coffee/Create
@@ -52,27 +92,56 @@ namespace CoffeeLand_UI.Areas.Admin.Controllers
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Create([Bind(Include = "ID,CoffeeName,Description,Price,AVGPoint,ExtraMaterialsID,CategoryID,ImageUrl,AltText")] Coffee coffee)
+        public ActionResult Create([Bind(Include = "ID,CoffeeName,Description,Price,AVGPoint,ImageUrl,AltText,IsPrepared,ExtraMaterialsID,CategoryID")] Coffee coffee)
         {
-            if (ModelState.IsValid)
-            {
-                _coffeeConcrete._coffeeRepository.Insert(coffee);
-                _coffeeConcrete._coffeeUnitOfWork.SaveChanges();
-                return RedirectToAction("Index");
-            }
+			Customer customer = Session["OnlineKullanici"] as Customer;
 
-            ViewBag.CategoryID = new SelectList(_categoryConcrete._categoryRepository.GetEntity(), "ID", "CategoryName", coffee.CategoryID);
-            ViewBag.ExtraMaterialsID = new SelectList(_extraMaterialsConcrete._extraMaterialRepository.GetEntity(), "ID", "Name", coffee.ExtraMaterialsID);
-            return View(coffee);
+			if (customer == null)
+			{
+				return Redirect("/Login/Login");
+			}
+			else if (customer.AuthorizationID == 1 || customer.AuthorizationID == 2)
+			{
+				if (ModelState.IsValid)
+				{
+					_coffeeConcrete._coffeeRepository.Insert(coffee);
+					_coffeeConcrete._coffeeUnitOfWork.SaveChanges();
+					return RedirectToAction("Index");
+				}
+
+				ViewBag.CategoryID = new SelectList(_categoryConcrete._categoryRepository.GetEntity(), "ID", "CategoryName", coffee.CategoryID);
+				ViewBag.ExtraMaterialsID = new SelectList(_categoryConcrete._categoryRepository.GetEntity(), "ID", "Name", coffee.ExtraMaterialsID);
+				return View(coffee);
+			}
+			else
+			{
+				return Redirect("/Coffee/Coffees");
+			}
+			
         }
 
         // GET: Admin/Coffee/Edit/5
         public ActionResult Edit(int id)
         {
-            Coffee coffee = _coffeeConcrete._coffeeRepository.GetById(id);
-            ViewBag.CategoryID = new SelectList(_categoryConcrete._categoryRepository.GetEntity(), "ID", "CategoryName", coffee.CategoryID);
-            ViewBag.ExtraMaterialsID = new SelectList(_extraMaterialsConcrete._extraMaterialRepository.GetEntity(), "ID", "Name", coffee.ExtraMaterialsID);
-            return View(coffee);
+			Customer customer = Session["OnlineKullanici"] as Customer;
+
+			if (customer == null)
+			{
+				return Redirect("/Login/Login");
+			}
+			else if (customer.AuthorizationID == 1 || customer.AuthorizationID == 2)
+			{
+				Coffee coffee = _coffeeConcrete._coffeeRepository.GetById(id);
+
+				ViewBag.CategoryID = new SelectList(_categoryConcrete._categoryRepository.GetEntity(), "ID", "CategoryName", coffee.CategoryID);
+				ViewBag.ExtraMaterialsID = new SelectList(_categoryConcrete._categoryRepository.GetEntity(), "ID", "Name", coffee.ExtraMaterialsID);
+				return View(coffee);
+			}
+			else
+			{
+				return Redirect("/Coffee/Coffees");
+			}
+			
         }
 
         // POST: Admin/Coffee/Edit/5
@@ -80,24 +149,50 @@ namespace CoffeeLand_UI.Areas.Admin.Controllers
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Include = "ID,CoffeeName,Description,Price,AVGPoint,ExtraMaterialsID,CategoryID,ImageUrl,AltText")] Coffee coffee)
+        public ActionResult Edit([Bind(Include = "ID,CoffeeName,Description,Price,AVGPoint,ImageUrl,AltText,IsPrepared,ExtraMaterialsID,CategoryID")] Coffee coffee)
         {
-            if (ModelState.IsValid)
-            {
-                _coffeeConcrete._coffeeRepository.Update(coffee);
-                _coffeeConcrete._coffeeUnitOfWork.SaveChanges();
-                return RedirectToAction("Index");
-            }
-            ViewBag.CategoryID = new SelectList(_categoryConcrete._categoryRepository.GetEntity(), "ID", "CategoryName", coffee.CategoryID);
-            ViewBag.ExtraMaterialsID = new SelectList(_extraMaterialsConcrete._extraMaterialRepository.GetEntity(), "ID", "Name", coffee.ExtraMaterialsID);
-            return View(coffee);
+			Customer customer = Session["OnlineKullanici"] as Customer;
+
+			if (customer == null)
+			{
+				return Redirect("/Login/Login");
+			}
+			else if (customer.AuthorizationID == 1 || customer.AuthorizationID == 2)
+			{
+				if (ModelState.IsValid)
+				{
+					_coffeeConcrete._coffeeRepository.Update(coffee);
+					_coffeeConcrete._coffeeUnitOfWork.SaveChanges();
+					return RedirectToAction("Index");
+				}
+				ViewBag.CategoryID = new SelectList(_categoryConcrete._categoryRepository.GetEntity(), "ID", "CategoryName", coffee.CategoryID);
+				ViewBag.ExtraMaterialsID = new SelectList(_categoryConcrete._categoryRepository.GetEntity(), "ID", "Name", coffee.ExtraMaterialsID);
+				return View(coffee);
+			}
+			else
+			{
+				return Redirect("/Coffee/Coffees");
+			}			
         }
 
         // GET: Admin/Coffee/Delete/5
         public ActionResult Delete(int id)
         {
-            Coffee coffee = _coffeeConcrete._coffeeRepository.GetById(id);
-            return View(coffee);
+			Customer customer = Session["OnlineKullanici"] as Customer;
+
+			if (customer == null)
+			{
+				return Redirect("/Login/Login");
+			}
+			else if (customer.AuthorizationID == 1 || customer.AuthorizationID == 2)
+			{
+				Coffee coffee = _coffeeConcrete._coffeeRepository.GetById(id);
+				return View(coffee);
+			}
+			else
+			{
+				return Redirect("/Coffee/Coffees");
+			}			
         }
 
         // POST: Admin/Coffee/Delete/5
@@ -105,17 +200,31 @@ namespace CoffeeLand_UI.Areas.Admin.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
-            Coffee coffee = _coffeeConcrete._coffeeRepository.GetById(id);
-            _coffeeConcrete._coffeeRepository.Delete(coffee);
-            _coffeeConcrete._coffeeUnitOfWork.SaveChanges();
-            return RedirectToAction("Index");
+			Customer customer = Session["OnlineKullanici"] as Customer;
+
+			if (customer == null)
+			{
+				return Redirect("/Login/Login");
+			}
+			else if (customer.AuthorizationID == 1 || customer.AuthorizationID == 2)
+			{
+				Coffee coffee = _coffeeConcrete._coffeeRepository.GetById(id);
+				_coffeeConcrete._coffeeRepository.Delete(coffee);
+				_coffeeConcrete._coffeeUnitOfWork.SaveChanges();
+				return RedirectToAction("Index");
+			}
+			else
+			{
+				return Redirect("/Coffee/Coffees");
+			}
+			
         }
 
         protected override void Dispose(bool disposing)
         {
             if (disposing)
             {
-                _coffeeConcrete._coffeeUnitOfWork.Dispose();
+				_coffeeConcrete._coffeeUnitOfWork.Dispose();
             }
             base.Dispose(disposing);
         }
